Revert "FRP: Prepare to remove randomisation of resolved IP addresses in getaddr()"
In the developers meeting July 22, 2010, a merge conflict between 'allmerged' and the
SVN BETA21 branch was discussed.  For more information see here:
<http://thread.gmane.org/gmane.network.openvpn.devel/3845>

It was decided on that meeting to remove this FRP process and let the improved getaddr_multi()
function from SVN BETA21 be included instead, knowing that this will remove the random resolve
feature without going through a Feature Removal Process.

This reverts commit 684790552dde0475745015a76762ecc5a11eedab.

Signed-off-by: David Sommerseth <dazo@users.sourceforge.net>